+++ changelog.md
@@ -1,5 +1,10 @@
 # Orientations.jl Change Log
 
+## Unreleased
+
+* Added indexing behavior on all types.
+* Added Dimensions extension.
+
 ## v1.0.1
 
 * Fixed ERP interpolation when `shortest_path = false`.